Block form of bind, deprecated
