+++ packages/ember-handlebars/tests/handlebars_test.js
@@ -526,162 +526,6 @@ test("View should update when the property used with the #with helper changes [D
   equal(view.$('#first').text(), "bazam", "view updates when a bound property changes");
 });
 
-test("should not update when a property is removed from the view", function() {
-  container.register('template:foo', EmberHandlebars.compile('<h1 id="first">{{#bind "view.content"}}{{#bind "foo"}}{{bind "baz"}}{{/bind}}{{/bind}}</h1>'));
-
-  view = EmberView.create({
-    container: container,
-    templateName: 'foo',
-
-    content: EmberObject.create({
-      foo: EmberObject.create({
-        baz: "unicorns"
-      })
-    })
-  });
-
-  appendView();
-
-  equal(view.$('#first').text(), "unicorns", "precond - renders the bound value");
-
-  var oldContent = get(view, 'content');
-
-  run(function() {
-    set(view, 'content', EmberObject.create({
-      foo: EmberObject.create({
-        baz: "ninjas"
-      })
-    }));
-  });
-
-  equal(view.$('#first').text(), 'ninjas', "updates to new content value");
-
-  run(function() {
-    set(oldContent, 'foo.baz', 'rockstars');
-  });
-
-  run(function() {
-    set(oldContent, 'foo.baz', 'ewoks');
-  });
-
-  equal(view.$('#first').text(), "ninjas", "does not update removed object");
-});
-
-test("Handlebars templates update properties if a content object changes", function() {
-  container.register('template:menu', EmberHandlebars.compile('<h1>Today\'s Menu</h1>{{#bind "view.coffee"}}<h2>{{color}} coffee</h2><span id="price">{{bind "price"}}</span>{{/bind}}'));
-
-  run(function() {
-    view = EmberView.create({
-      container: container,
-      templateName: 'menu',
-
-      coffee: EmberObject.create({
-        color: 'brown',
-        price: '$4'
-      })
-    });
-  });
-
-  appendView();
-
-  equal(view.$('h2').text(), "brown coffee", "precond - renders color correctly");
-  equal(view.$('#price').text(), '$4', "precond - renders price correctly");
-
-  run(function() {
-    set(view, 'coffee', EmberObject.create({
-      color: "mauve",
-      price: "$4.50"
-    }));
-  });
-
-  equal(view.$('h2').text(), "mauve coffee", "should update name field when content changes");
-  equal(view.$('#price').text(), "$4.50", "should update price field when content changes");
-
-  run(function() {
-    set(view, 'coffee', EmberObject.create({
-      color: "mauve",
-      price: "$5.50"
-    }));
-  });
-
-  equal(view.$('h2').text(), "mauve coffee", "should update name field when content changes");
-  equal(view.$('#price').text(), "$5.50", "should update price field when content changes");
-
-  run(function() {
-    set(view, 'coffee.price', "$5");
-  });
-
-  equal(view.$('#price').text(), "$5", "should update price field when price property is changed");
-
-  run(function() {
-    view.destroy();
-  });
-});
-
-test("Template updates correctly if a path is passed to the bind helper", function() {
-  container.register('template:menu', EmberHandlebars.compile('<h1>{{bind "view.coffee.price"}}</h1>'));
-
-  view = EmberView.create({
-    container: container,
-    templateName: 'menu',
-
-    coffee: EmberObject.create({
-      price: '$4'
-    })
-  });
-
-  appendView();
-
-  equal(view.$('h1').text(), "$4", "precond - renders price");
-
-  run(function() {
-    set(view, 'coffee.price', "$5");
-  });
-
-  equal(view.$('h1').text(), "$5", "updates when property changes");
-
-  run(function() {
-    set(view, 'coffee', { price: "$6" });
-  });
-
-  equal(view.$('h1').text(), "$6", "updates when parent property changes");
-});
-
-test("Template updates correctly if a path is passed to the bind helper and the context object is an ObjectController", function() {
-  container.register('template:menu', EmberHandlebars.compile('<h1>{{bind "view.coffee.price"}}</h1>'));
-
-  var controller = ObjectController.create();
-
-  var realObject = EmberObject.create({
-    price: "$4"
-  });
-
-  set(controller, 'model', realObject);
-
-  view = EmberView.create({
-    container: container,
-    templateName: 'menu',
-
-    coffee: controller
-  });
-
-  appendView();
-
-  equal(view.$('h1').text(), "$4", "precond - renders price");
-
-  run(function() {
-    set(realObject, 'price', "$5");
-  });
-
-  equal(view.$('h1').text(), "$5", "updates when property is set on real object");
-
-  run(function() {
-    set(controller, 'price', "$6" );
-  });
-
-  equal(view.$('h1').text(), "$6", "updates when property is set on object controller");
-});
-
 test("should update the block when object passed to #if helper changes", function() {
   container.register('template:menu', EmberHandlebars.compile('<h1>{{#if view.inception}}{{view.INCEPTION}}{{/if}}</h1>'));
